test(renderer): cover Cyrillic design text fallback
